Create module for getting baseUrl
