+++ react/SearchAction.tsx
@@ -1,35 +1,14 @@
 import React, { FC } from 'react'
 import { Helmet } from 'react-helmet'
-import { canUseDOM } from 'vtex.render-runtime'
 
-interface Runtime {
-  rootPath?: string
-}
+import { getBaseUrl } from './modules/baseUrl'
 
 interface Props {
   searchTermPath?: string
 }
 
-declare let global: {
-  __hostname__: string
-  __protocol__: string
-  __pathname__: string
-  __RUNTIME__: Runtime
-}
-
-declare let window: {
-  __RUNTIME__: Runtime
-  location: Location
-}
-
 const SearchAction: FC<Props> = ({ searchTermPath }) => {
-  const protocol = 'https'
-  const hostname = canUseDOM ? window.location.hostname : global.__hostname__
-  const rootPath = canUseDOM
-    ? window.__RUNTIME__.rootPath
-    : global.__RUNTIME__.rootPath
-
-  const baseUrl = `${protocol}://${hostname}${rootPath ?? ''}`
+  const baseUrl = getBaseUrl()
   const path = !searchTermPath ? '/' : searchTermPath
 
   const schema = {
