+++ react/tsconfig.json
@@ -4,6 +4,7 @@
     "noEmitOnError": false,
     "typeRoots": ["node_modules/@types"],
     "types": ["node", "jest", "graphql"],
+    "lib": ["dom"],
     "module": "esnext",
     "moduleResolution": "node",
     "target": "es2017"
